% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parenv.r
\name{as.envlist}
\alias{as.envlist}
\title{Convert a list of environments to an \code{envlist} object.}
\usage{
as.envlist(x)
}
\arguments{
\item{x}{A list of environments.}
}
\description{
Convert a list of environments to an \code{envlist} object.
}
\keyword{internal}

