% Generated by roxygen2 (4.0.1): do not edit by hand
\name{enclosing_env}
\alias{enclosing_env}
\title{Find the environment that encloses of a function.}
\usage{
enclosing_env(f)
}
\arguments{
\item{f}{The name of a function.}
}
\description{
This is a wrapper around \code{\link{environment}} with a
consistent syntax.
}
\examples{
enclosing_env("plot")
enclosing_env("t.test")
}

