% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit-promise.R
\name{explicit}
\alias{explicit}
\alias{eval2}
\title{Tools for making promises explicit}
\usage{
explicit(x)

eval2(x, data = NULL, env = parent.frame())
}
\arguments{
\item{x}{expression to make explicit, or to evaluate.}

\item{data}{Data in which to evaluate code}

\item{env}{Enclosing environment to use if data is a list or data frame.}
}
\description{
Deprecated: please use the lazyeval package instead.
}
