% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.r
\name{is_promise}
\alias{is_promise}
\alias{promise_info}
\title{Promise info}
\usage{
is_promise(x)

promise_info(x)
}
\arguments{
\item{x}{unquoted object name}
}
\description{
Promise info
}
\examples{
x <- 10
is_promise(x)
(function(x) is_promise(x))(x = 10)
}
\seealso{
Other promise tools: 
\code{\link{uneval}()}
}
\concept{promise tools}
