% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{mem_change}
\alias{mem_change}
\title{Determine change in memory from running code}
\usage{
mem_change(code)
}
\arguments{
\item{code}{Code to evaluate.}
}
\value{
Change in memory (in megabytes) before and after running code.
}
\description{
Determine change in memory from running code
}
\examples{
# Need about 4 mb to store 1 million integers
mem_change(x <- 1:1e6)
# We get that memory back when we delete it
mem_change(rm(x))
}
