% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mem_used}
\alias{mem_used}
\title{How much memory is currently used by R?}
\usage{
mem_used()
}
\value{
Megabytes of ram used by R objects.
}
\description{
R breaks down memory usage into Vcells (memory used by vectors) and
Ncells (memory used by everything else). However, neither this distinction
nor the "gc trigger" and "max used" columns are typically important. What
we're usually most interested in is the the first column: the total memory
used. This function wraps around \code{gc()} to return the total amount of
memory (in megabytes) currently used by R.
}
\examples{
mem_used()
}

