% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_c.R
\name{names_c}
\alias{names_c}
\title{Extract function table from names.c from R subversion repository.}
\usage{
names_c()
}
\value{
A data frame with columns
\item{name}{the function name in R}
\item{c-entry}{The name of the corresponding C function, actually declared
  in ../include/Internal.h. All start with "do_", return SEXP, and
  have argument list (SEXP call, SEXP op, SEXP args, SEXP env)}
\item{offset}{the 'op' (offset pointer) above; used for C functions
	 which deal with more than one R function}
\item{eval}{XYZ (three digits) \cr
 \cr
 X=0 says that we should force R_Visible on \cr
 X=1 says that we should force R_Visible off \cr
 X=2 says that we should switch R_Visible on but let the C code update it. \cr
 \cr
 Y=1 says that this is an internal function which must
     be accessed with a	.Internal(.) call, any other value is
     accessible directly and printed in R as ".Primitive(..)".\cr
 \cr
 Z=0 says don't evaluate (SPECIALSXP).\cr
 Z=1 says evaluate arguments before calling (BUILTINSXP)}
\item{arity}{How many arguments are required/allowed;  "-1"	meaning ``any''}
\item{pp-kind}{Deparsing Info (-> PPkind in ../include/Defn.h )}
\item{precedence}{Operator precedence (-> PPprec in ../include/Defn.h )}
\item{rightassoc}{Right or left associative operator}
}
\description{
Since this is an expensive operation, it is done once and cached within
a session.
}
\keyword{internal}
