% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show_c_source}
\alias{show_c_source}
\title{Find C source code for internal R functions}
\usage{
show_c_source(fun)
}
\arguments{
\item{fun}{.Internal or .Primitive function call.}
}
\description{
Opens a link to code search on github.
}
\examples{
\donttest{
show_c_source(.Internal(mean(x)))
show_c_source(.Primitive(sum(x)))
}
}

