\name{track_copy}
\alias{track_copy}
\title{Track if an object is copied}
\usage{
  track_copy(var, env = parent.frame(), quiet = FALSE)
}
\description{
  The title is somewhat misleading: rather than checking if
  an object is modified, this really checks to see if a
  name points to the same object.
}

