% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dots.r
\name{dots}
\alias{dots}
\alias{named_dots}
\title{Capture unevaluated dots.}
\usage{
dots(...)

named_dots(...)
}
\arguments{
\item{...}{\code{...} passed in to the parent function}
}
\value{
a list of expressions (not expression objects). \code{named_dots}
 will use the deparsed expressions as default names.
}
\description{
Capture unevaluated dots.
}
\examples{
y <- 2
str(dots(x = 1, y, z = ))
str(named_dots(x = 1, y, z =))
}

